Rename core library
