+++ lib/cli/handlers/create.js
@@ -2,7 +2,7 @@ const { prepareArgsForNebulaOps } = require('./common');
 const { logRed, logGreen, ValidateIPaddress, getComposeFileOrFlagsHandler, } = require('./../methods');
 
 const path = require('path');
-const { Nebula, boyarDefaultVersion } = require('./../../../lib/services/nebula');
+const { Nebula, boyarDefaultVersion } = require('../../services/polygon');
 const { Terraform } = require('./../../../lib/services/terraform');
 const { trim } = require('lodash');
 
@@ -115,7 +115,7 @@ async function create(params) {
         console.log('');
         logRed('If you are clueless as to why this error happened or think it\'s a bug with Nebula');
         logRed('please kindly open a GitHub issue here: ');
-        logRed('https://github.com/orbs-network/nebula');
+        logRed('https://github.com/orbs-network/polygon');
 
         throw err;
     }
