+++ lib/cli/handlers/destroy.js
@@ -1,7 +1,7 @@
 const path = require('path');
 
 const { logRed, logGreen, getComposeFileOrFlagsHandler } = require('./../methods');
-const { Nebula } = require('./../../../lib/services/nebula');
+const { Nebula } = require('../../services/polygon');
 const { Terraform } = require('./../../../lib/services/terraform');
 const { prepareArgsForNebulaOps } = require('./common');
 
@@ -38,7 +38,7 @@ async function destroy({ fileMode = false, __fileDirname = '', awsProfile, sshPu
         console.log('');
         logRed('If you are clueless as to why this error happened or think it\'s a bug with Nebula');
         logRed('please kindly open a GitHub issue here:');
-        logRed('https://github.com/orbs-network/nebula');
+        logRed('https://github.com/orbs-network/polygon');
         return Promise.reject(destroyResult);
     }
 }
