+++ lib/cli/handlers/update.js
@@ -17,7 +17,7 @@ function _update({ name, region, configPath: _configPath, chainVersion, awsProfi
         console.log('');
         logRed('If you are clueless as to why this error happened or think it\'s a bug with Nebula');
         logRed('please kindly open a GitHub issue here:');
-        logRed('https://github.com/orbs-network/nebula');
+        logRed('https://github.com/orbs-network/polygon');
 
         throw new Error('No config path specified!');
     }
@@ -62,7 +62,7 @@ async function update(params) {
         console.log('');
         logRed('If you are clueless as to why this error happened or think it\'s a bug with Nebula');
         logRed('please kindly open a GitHub issue here:');
-        logRed('https://github.com/orbs-network/nebula');
+        logRed('https://github.com/orbs-network/polygon');
 
         throw execReturnValue.stderr;
     }
